Add 'blocked on $direction' to the s2n I/O APIs
This change takes s2n prior 'more' parameter and recasts it as a new 'blocked'
parameter with 3 potential values:

    S2N_NOT_BLOCKED
    S2N_BLOCKED_ON_READ
    S2N_BLOCKED_ON_WRITE

which indicate how s2n is currently blocked. This permits non-blocking callers
to detect when they should call s2n again. For example if s2n is in the
S2N_BLOCKED_ON_READ status, then the caller can wait until it detects that
there is pending I/O available to be read on the underlying socket.