Add specification of SHA1
The includes the functional version of SHA1 that we had in the examples
directory of the Cryptol repository, as well as a more imperative
version (written in Cryptol) that consists of initialization, update,
and finalization functions. The SAWScript file proves that the
imperative version is equivalent to the functional version for various
specific message sizes.

It turns out that ABC is remarkably efficient at proving this! Probably
because the block function is completely identical on both sides of the
equality.

Note that I've included SHA1 simply as an example hash algorithm since
a clean specification of it was readily at hand. The same approach to
creating an imperative version should work equally well for any other
hash algorithm.