Start abstracting away from SHA256.
Not complete yet. We replaced the SHA256-specific hash calls with
calls to generic `hash_init_c_state`, `hash_update_c_state`, and
`hash_digest_c_state`, which we define as `undefined` and
uninterpret. This allows us to eliminate most SHA256 specific code
from the spec, but we now need to deal with constraining values in the
HMAC state enough to allow the simulation to complete without error.

We are not verifying the `s2n_hash_*` function specs against a real
hash implemention right now, but the idea is to write them in such a
way that they would be verifiable.