Add support for variable prf hash
RFC5246(TLS1.2) allows new cipher suites to define their own PRF and
associated hash function(if applicable)[1]. All of the cipher suites we
support use a P_hash based PRF, so we only need to specify the hash algorithm.

Protocol versions < TLS1.2 use a hardcoded PRF. What if we negotiate TLS1.2
and select a cipher suite defined before TLS1.2? Ex: TLS1.2 with
TLS_RSA_WITH_3DES_EDE_CBC_SHA selected. RFC5246 states that "This PRF with the
SHA-256 hash function is used for all cipher suites defined in this document
and in TLS documents published prior to this document when TLS 1.2 is negotiated."

[1] https://tools.ietf.org/html/rfc5246#section-5