Add a basic handshake integration test
This change adds a handshake integration test via an Openssl based
Python client. The intention is to have this test run against s2nd
in TravisCI. That way we'll get feedback on handshake regression as
changes are added to s2n. Any time new cipher suites are added to s2n,
we should also add them to s2n_handshake_test.py. In the future, we could
have it test different combinations of cipher suite and protocol version.

It would be tricky to make this a build requirement because
of the dependencies required(Openssl, Python > 2.7.10).