Settle for a higher proto cipher if needed
This change will have the server fallback to a cipher match that
the client shouldn't support according to their protocol version
if we have no other options. For example, if the only cipher suites a
TLSv1.0 client matches with the server have minimum proto version
TLSv1.2, lets try to negotiate one of those TLSv1.2 ciphers instead
of failing the handshake.