Make extensions tests portable
s2n could end up blocking either on read or write in a test case,
depending on how much space is available in a pipe buffer. On my
mac, the pipe fills and s2n is blocked on read. On our travis build
hosts it's the other way around.

This change makes the check portable.