Return empty renegotiation_info extension if applicable
This change fixes an interoperability problem between s2n in
server mode and tls clients that implement the secure renegotiation
extension. This change does not enable renegotiation in s2n.
An empty renegotion_info extension is now sent in the ServerHello if
the client includes either:
1. TLS_EMPTY_RENEGOTIATION_INFO_SCSV in its cipher suite list
2. An empty renegotiation_info extension

RFC5746 suggests that TLS clients MAY terminate the handshake if the
server does not send an empty renegotiation_info extension in response
to a client's [1]. The RFC states that servers that don't support renegotiation
at all(like s2n), MUST implement the minimal version of the extension for
the initial handshake [2].

[1] https://tools.ietf.org/html/rfc5746#section-4.1
[2] https://tools.ietf.org/html/rfc5746#section-4.3